pyproject.toml
[project]
name = "karldbot"
version = "0.2.0"
description = "Karl the coder - Multi-agent RL system for data science code generation"
authors = [{ name = "Flávio Codeço Coelho", email = "fccoelho@gmail.com" }]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "GPLv3" }
dependencies = [
    "ruff>=0.6.5,<0.7",
    "base-ai-agent",
    "pydantic>=2.9.2,<3",
    "duckdb>=1.1.0,<2",
    "pandas>=2.2.2,<3",
    "python-dotenv>=1.0.1,<2",
    "jinja2>=3.1.4,<4",
    "fire",
    "pyaml>=24.7.0,<25",
    "tqdm>=4.66.5,<5",
    "matplotlib>=3.9.2,<4",
    "loguru",
    "tabulate>=0.9.0,<0.10",
    "numpy>=1.26.0,<2",
]

[project.scripts]
karl = "karldbot.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.3.3,<9",
    "pytest-cov>=5.0.0,<6",
    "pytest-ruff>=0.4.1,<0.5",
    "pytest-asyncio>=0.24.0,<0.25",
    "mypy>=1.11.0,<2",
]

[tool.uv.sources]
base-ai-agent = { git = "https://github.com/Deeplearn-PeD/base-ai-agent.git" }

[tool.hatch.build.targets.sdist]
include = ["./karldbot"]

[tool.hatch.build.targets.wheel]
include = ["./karldbot"]

[tool.hatch.build.targets.wheel.sources]
"./karldbot" = "karldbot"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
